global script added in _app.tsx
